refactor: 클래스명, 필드 타입 수정
- 기존 StudyCafe 엔티티의 DTO 객체인 StudyCafeDto를 쓰이는 용도를 명확히 하기 위해 StudyCafeAddDto로 이름 변경
- Owner 필드를 OwnerDto 로 수정하여 엔티티 대신 Dto를 필드로 가지게 수정
